some changes

+++ 1/WinTerm.h
@@ -1,6 +1,9 @@
 #pragma once
+
 #include <Windows.h>
 
 TCHAR getch();
+
 void clear_screen();
-void gotoxy(int x, int y);
+
+void gotoxy(short x, short y);
